+++ Cargo.toml
@@ -60,7 +60,7 @@ dialoguer = "0.11"
 # Text processing
 regex = "1"
 text-processing-rs = "0.2.2"
-similar = { version = "2", features = ["inline"] }
+similar = { version = "2", features = ["bytes", "inline", "unicode", "wasm32_web_time"] }
 owo-colors = { version = "4", features = ["supports-colors"] }
 time = { version = "0.3", features = ["formatting", "local-offset", "macros"] }
 comfy-table = { version = "7.2", default-features = true, features = ["tty"] }
